+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Support for SYSVER system variable
 - CICS preprocessor now recognizes DFHVALUE constants
 - Enhanced commands for continuation handling and trimming of oversized lines
+- Provide the name of a missing variable or ordinary symbol in messages
 
 #### Fixed
 - Incorrect attribute values generated when literals are substituted in CA expressions